requirements.txt
# --------- pytorch --------- #
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.10.1+cpu
torchvision==0.11.2+cpu
timm==0.6.7
pytorch-lightning==1.7.1
torchmetrics==0.9.3

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- demo -----------#
gradio==3.3.1

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
setuptools==59.5.0